+++ rocketmq-remoting/src/protocol/admin/topic_stats_table.rs
@@ -35,8 +35,16 @@ impl TopicStatsTable {
         }
     }
 
-    pub fn get_offset_table(&self) -> HashMap<MessageQueue, TopicOffset> {
-        self.offset_table.clone()
+    pub fn get_offset_table(&self) -> &HashMap<MessageQueue, TopicOffset> {
+        &self.offset_table
+    }
+
+    pub fn get_offset_table_mut(&mut self) -> &mut HashMap<MessageQueue, TopicOffset> {
+        &mut self.offset_table
+    }
+
+    pub fn into_offset_table(self) -> HashMap<MessageQueue, TopicOffset> {
+        self.offset_table
     }
 
     pub fn set_offset_table(&mut self, offset_table: HashMap<MessageQueue, TopicOffset>) {